refactor: 이름이 모호하던 ResponseHandler를 RequestHandler로 변경 - 기존 RequestHandler는 HTTP 파싱만 진행하는 것으로 보여 HttpParser를 바로 사용한다.
