+++ src/main/java/codesquad/WebServer.java
@@ -3,7 +3,6 @@ package codesquad;
 import codesquad.webserver.HttpTask;
 import codesquad.webserver.HttpThreadPool;
 import codesquad.webserver.RequestHandler;
-import codesquad.webserver.ResponseHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -16,7 +15,6 @@ public class WebServer {
     private static final Logger logger = LoggerFactory.getLogger(WebServer.class);
     private static final int THREAD_POOL_SIZE = 10;
     private RequestHandler requestHandler;
-    private ResponseHandler responseHandler;
     private ServerSocket serverSocket;
     private HttpThreadPool httpThreadPool;
 
@@ -24,7 +22,6 @@ public class WebServer {
         serverSocket = new ServerSocket(port);
         logger.debug("Listening for connection on port {} ....", port);
         requestHandler = new RequestHandler();
-        responseHandler = new ResponseHandler();
         httpThreadPool = new HttpThreadPool(Executors.newFixedThreadPool(THREAD_POOL_SIZE));
     }
 
@@ -32,7 +29,7 @@ public class WebServer {
         while (true) { // 무한 루프를 돌며 클라이언트의 연결을 기다립니다.
             final Socket clientSocket = serverSocket.accept();
             logger.debug("Client connected {} {}", clientSocket.getLocalPort(), clientSocket.getLocalAddress());
-            httpThreadPool.execute(new HttpTask(clientSocket, requestHandler, responseHandler));
+            httpThreadPool.execute(new HttpTask(clientSocket, requestHandler));
         }
     }
 }
